requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
decorator==5.1.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-taggit==4.0.0
executing==2.0.0
hiredis==2.2.3
iniconfig==2.0.0
ipython==8.16.1
jedi==0.19.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.16.1
python-decouple==3.8
python-json-logger==2.0.7
rich==13.6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
wcwidth==0.2.8
